requirements.txt
fastapi
uvicorn
requests
python-multipart
pydantic
PyPDF2
python-dotenv